+++ src/containers/gameController.js
@@ -11,10 +11,31 @@ export default {
     // The game loop should set up a new game by creating Players and Gameboards.
     // 1. Create gameboards
     // 2. Create players and pass in their gameboard and the opponent's gameboard.
-    const playerOneBoard = Gameboard();
-    const playerTwoBoard = Gameboard();
-    this.playerOne = pipe(Player, isHuman)(playerOneBoard, playerTwoBoard);
-    this.playerTwo = pipe(Player, isHuman)(playerTwoBoard, playerOneBoard);
+    //  Do I only need to pass the opponent's board?
+    this.playerOneBoard = Gameboard();
+    this.playerTwoBoard = Gameboard();
+    this.playerOne = pipe(Player, isHuman)(this.playerOneBoard, this.playerTwoBoard);
+    this.playerTwo = pipe(Player, isHuman)(this.playerTwoBoard, this.playerOneBoard);
+
+    this.playerOneBoard.placeShip([2, 2], false);
+    this.playerTwoBoard.placeShip([6, 2], false);
+    this.switchPlayers();
+    console.table(this.playerTwo.board.board);
   },
-  playRound() {},
+  switchPlayers(player) {
+    if (player) {
+      // Looking into Lodash _.isEqual()
+      // Could add a turn property to player object that takes a boolean
+      this.activePlayer = player === this.playerOne ? this.playerTwo : this.playerOne;
+    } else {
+      // Initially set a random player as the active player
+      const players = [this.playerOne, this.playerTwo];
+      this.activePlayer = players[Math.floor(Math.random() * 2)];
+    }
+  },
+  playRound(coordinate) {
+    this.activePlayer.attack(coordinate);
+    this.switchPlayers(this.activePlayer);
+    this.activePlayer.attack(coordinate);
+  },
 };