+++ src/testing/gameController.test.js
@@ -6,8 +6,6 @@ const battleship = gameController;
 
 beforeEach(() => {
   battleship.init();
-  battleship.playerOne.board.placeShip([2, 2], false);
-  battleship.playerTwo.board.placeShip([6, 2], false);
 });
 
 const playerOneShipCoordinates = [
@@ -42,12 +40,34 @@ describe(`Tests if gameController.playRound works`, () => {
 
 describe(`Tests if players can attack one another`, () => {
   test(`Player two attacks and sinks player one's ship`, () => {
-    const shipCoordinates = [
-      [2, 2],
-      [3, 2],
-      [4, 2],
-    ];
-    shipCoordinates.forEach(battleship.playerTwo.attack);
-    expect(battleship.playerOne.board.board[8][1].isSunk()).toBeTruthy();
+    const playerOne = battleship.playerOne;
+    const playerTwo = battleship.playerTwo;
+    if (playerOne.attack.toString() === playerTwo.attack.toString()) {
+      // playerOne and playerTwo are humans
+      const shipCoordinates = [
+        [2, 2],
+        [3, 2],
+        [4, 2],
+      ];
+      shipCoordinates.forEach((coordinate) => {
+        battleship.playerTwo.attack(coordinate);
+      });
+    } else {
+      let i = 0;
+      while (i < 100) {
+        battleship.playerTwo.attack();
+        i += 1;
+      }
+    }
+
+    expect(battleship.playerOneBoard.board[8][1].isSunk()).toBeTruthy();
+  });
+});
+
+describe(`Tests if a round can be played`, () => {
+  test(`The active player should be different after the round is done`, () => {
+    const initalActivePlayer = battleship.activePlayer;
+    battleship.playRound([2, 2]);
+    expect(battleship.activePlayer === initalActivePlayer).toBeFalsy();
   });
 });